add prettier/eslint and run formatting
