+++ .github/workflows/production.yml
@@ -6,7 +6,7 @@ name: UpdateProdPipeline
 on:
   # Triggers the workflow on push or pull request events but only for the main branch
   push:
-    branches: [ main ]
+    branches: [main]
 
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
@@ -20,7 +20,6 @@ jobs:
 
     # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
-    
       # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
       - name: Check out and update pipeline repository
         uses: actions/checkout@master
@@ -42,7 +41,7 @@ jobs:
         run: |
           git checkout main
           git pull origin main
-     
+
       # Update website
       - name: Update Website
         run: |
@@ -50,5 +49,3 @@ jobs:
           currentDate=`date +"%Y-%m-%d %T"`
           git commit -m "Website Updated by GitHub Actions $currentDate"
           git push -u origin main
-          
-
