+++ .prettierignore
@@ -1,10 +1,23 @@
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+
+# dependencies
+/node_modules
+/.pnp
+.pnp.js
+
 # testing
-coverage
+/coverage
 
 # production
-build
-public
+/build
 
-# artifacts
-node_modules
-assets
+# misc
+.DS_Store
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
