+++ src/components/Landing/landing.css
@@ -1,84 +1,88 @@
 .header-img-container {
-    height: 50vh;
+  height: 50vh;
 }
-#bkg{
+#bkg {
+  width: 100%;
+  z-index: -1;
+}
+#pfp {
+  position: relative;
+  height: 30vh;
+  z-index: 0;
+  margin: 0 auto;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
+.content_text {
+  position: relative;
+  top: -15vh;
+}
+@media (min-width: 1000px) {
+  #content {
+    margin: 0px 10% 0px 10%;
+  }
+
+  .modal-box {
+    width: 50%;
+  }
+}
+@media (max-width: 999px) {
+  #content {
+    margin: 0px 0% 0px 0%;
+  }
+
+  .modal-box {
     width: 100%;
-    z-index: -1;
+  }
 }
-#pfp{
-    position: relative;
-    height: 30vh;
-    z-index: 0;
-    margin: 0 auto;
-    left: 50%;
-    transform: translate(-50%, -50%);
+.header_text {
+  font-family: "Arvo";
 }
-.content_text{
-    position: relative;
-    top: -15vh;
+.body_text {
+  font-family: "Open Sans";
 }
-@media(min-width: 1000px){
-    #content{
-        margin: 0px 10% 0px 10%;   
-    }
-
-	.modal-box{
-		width: 50%;
-	}
+.content_text div {
+  padding: 1%;
 }
-@media(max-width: 999px){
-    #content{
-        margin: 0px 0% 0px 0%;   
-    }
-
-	.modal-box{
-		width: 100%;
-	}
+h4,
+p {
+  padding: 1%;
 }
-.header_text{
-    font-family: "Arvo";
+#summary p,
+#projects p,
+.card_small p {
+  margin: 0px;
 }
-.body_text{
-    font-family: "Open Sans";
+.card_img img {
+  width: 100%;
 }
-.content_text div{
-    padding: 1%;
+h5,
+.card_small h6 {
+  margin-top: 0px;
 }
-h4, p{
-    padding: 1%;
+.card_small p {
+  padding: 0px;
 }
-#summary p, #projects p, .card_small p{
-    margin: 0px;
+.card_small {
+  padding: 2%;
 }
-.card_img img{
-    width: 100%;
+.expandPrompt {
+  padding: 0px;
+  margin: 0px;
 }
-h5, .card_small h6{
-    margin-top: 0px;
-}
-.card_small p{
-    padding:0px;
-}
-.card_small{
-    padding:2%;
-}
-.expandPrompt{
-    padding: 0px;
-    margin: 0px;
-}
-.card_content img{
-	width: 100%;
+.card_content img {
+  width: 100%;
 }
 
-.modal-content{
-	padding: 24px;
+.modal-content {
+  padding: 24px;
 }
 
-.modal-content img{
-	width: 100%;
+.modal-content img {
+  width: 100%;
 }
 
-.skillset li{
-	margin: 0;
-	padding: 0;
+.skillset li {
+  margin: 0;
+  padding: 0;
 }
